fr: fix translation of "check"
"Check" was being translated literally as "put a checkmark on", which made no sense at all. The original meaning seems to be to "make a list of the questions you find interesting" for the first case, and "remove from the list the questions whose answers you can find online" for the other.

Also, fixed the translation of "Definitely don't try to ask everything from the list", which had a weird use of "demander", taking the translation of ask too literally; and added the missing "(Respect the interviewer's time and show initiative by finding answers on your own if they're already published)".